GJ-76: Remove deprecating functionality and add more deprecation warnings
(cherry picked from commit 55b7c95d345f8c74bff73973285f82a3e17296eb)

cleanup

Respond to review comments

Typo

+++ src/pygribjump/pygribjump_type.py
@@ -134,7 +134,8 @@ class ExtractionRequest:
         """
         total = sum(self._shape)
         idx_iter = (idx for (low, high) in self.ranges for idx in range(low, high))
-        return np.fromiter(idx_iter, int, count=total)
+        # Use an explicit dtype: plain `int` maps to a platform-dependent width.
+        return np.fromiter(idx_iter, np.int64, count=total)
 
     def __repr__(self) -> str:
         return repr(self._request)
